Use environ contextmanager, ripped from test utils and promoted to the big leagues
