+++ python_modules/dagster/dagster/_core/instance_for_test.py
@@ -8,31 +8,9 @@ import yaml
 
 from dagster._utils.error import serializable_error_info_from_exc_info
 
+from .._utils.env import environ
 from .._utils.merger import merge_dicts
 from .instance import DagsterInstance
-
-
-@contextmanager
-def environ(env: Mapping[str, str]) -> Iterator[None]:
-    """Temporarily set environment variables inside the context manager and
-    fully restore previous environment afterwards.
-    """
-    previous_values = {key: os.getenv(key) for key in env}
-    for key, value in env.items():
-        if value is None:
-            if key in os.environ:
-                del os.environ[key]
-        else:
-            os.environ[key] = value
-    try:
-        yield
-    finally:
-        for key, value in previous_values.items():
-            if value is None:
-                if key in os.environ:
-                    del os.environ[key]
-            else:
-                os.environ[key] = value
 
 
 @contextmanager
