+++ python_modules/dagster/dagster/_utils/env.py
@@ -1,0 +1,29 @@
+import os
+from contextlib import contextmanager
+from typing import (
+    Iterator,
+    Mapping,
+)
+
+
+@contextmanager
+def environ(env: Mapping[str, str]) -> Iterator[None]:
+    """Temporarily set environment variables inside the context manager and
+    fully restore previous environment afterwards.
+    """
+    previous_values = {key: os.getenv(key) for key in env}
+    for key, value in env.items():
+        if value is None:
+            if key in os.environ:
+                del os.environ[key]
+        else:
+            os.environ[key] = value
+    try:
+        yield
+    finally:
+        for key, value in previous_values.items():
+            if value is None:
+                if key in os.environ:
+                    del os.environ[key]
+            else:
+                os.environ[key] = value
